Set the leader in vimrc.
This is related to #3, the mappings that were working without explicit
sourcing were the ones that did not utilize the `leader` key.
`:scriptnames` gives the order in which vim loads all the scripts and
according to that, `keybindings.vim` is loaded before `settings.vim`.
The `leader` is defined in `settings.vim` as a result all keybindings
using the `leader` were not working.
My guess as to why it worked is that the function that is doing the
loading is probably changing the order i.e. perhaps `settings.vim` is
being loaded first (this is a maybe, according to what I understand
after reading the code it should load the scripts in the plugin
directory top to bottom).
Regardless, setting the `leader` before `keybindings.vim` i.e. in
`vimrc` works!