Restructure how files are arranged.
This commit brings in two changes.

1. Highlights have been moved into their own file.
They have been placed under the `/after/plugin` directory. This is
because we want to load these settings after the colorscheme has loaded.
2. Settings for plugins have been split up into individual files under
`/plugin`. I wanted them to be under `/after/plugin` but it seems that
some plugins need these settings to be set prior to being loaded (this
is an implementation detail the developer chose) hence I have no other
option but to place them inside `/plugin`.