Enable folding by syntax.
The commit also introduces a handy key binding i.e. Tab is mapped to
`za` which toggles folds. From the perspective of the terminal, `Tab`
and `Ctrl-i` are the same keys so this keybinding prevents me from using
the `Ctrl-i` binding (which defaults to move forward in the jump list).
However, since I don't use the jump list that often, this trade off is
acceptable to me.