Add a file preview.
Picked it up from fzf author's vimrc (actually found it in a YouTube
video and that guy had picked it up from the fzf author however I seem
to have lost the link to the video). Essentially, this change adds a
window on the left of the split when the `:Files` command is issued and
shows a live preview of the file under the cursor.