A (unoptimized) wrapper around base16-shell
base16-shell does not provide a convenient way to switch color schemes on
the fish shell. This commit introduces a new function =colors= that accepts
a single argument denoting the name of the scheme that you wish to set.
Simple logic exists to check the validity of the argument, helpful comments
are printed.

The _list_colors function is called several times, ideally we would want to
cache the results of this function the first time so that subsequent calls
are faster.