deploy: init deploy script
`deploy` is the main script that will expose the api to the dotfile
system. To begin with, we need to have a book keeping system for topics
that are currently enabled and an interface to add/remove topics from
this system. This commit introduces two (private) methods:

+ `_enable_topic`: adds topic to ENABLED_TOPICS if it doesn't already
exist

+ `_diable_topic`: removes topic from ENABLED_TOPICS if isn't already
disabled

A universal variable ENABLED_TOPICS is used here since we need to
persist this data throughout sessions and reboots.