deploy: complete _link_topics
_link_topics accepts TOPICS and symlinks all files (that are not `_init`,
`bin` or `*.md`) under the TOPIC to $XDG_CONFIG_HOME/$TOPIC. Under the
hood, this function called `_mklink` which accepts a `src` and a
`target`. the `-f` flag is passed to `ln` so that the function can be
run multiple times. `_mklink` also creates $XDG_CONFIG_HOME/$TOPIC if it
doesn't exist.