fish/functions/g: init a wrapper around git
git status is one of the most frequent commands that I run, as such
optimizing this process is a no brainer. this commit introduces a
function g which wraps around the git command. when called without any
args, git status is called. if args are passed then the command is
delegated to git.