base: set abbreviations to global scope
by default, fish sets the scope of all abbreviations to universal. this
results in a dirty uninstall of topics (abbreviations of topics are left
behind if not manually removed). the fix is to set
fish_user_abbreviations to the global scope before the first
abbreviation.

fish evaluates files under conf.d in alphabetical order. as such,
base.fish is the first file that is evaulated by fish.