text file to track enabled topics
using long universal variables affects the performance of fish
drastically, this commit replaces the existing system of tracking
enabled topics ie universal variable ENABLED_TOPICS with a file
enabled_topics. The file is ignored from version control and is hidden
with a "dot" since it is not meant to be edited directly by the user.