kill _mk_link
vastly simplify the linking process! the custom linking strategy was
poorly designed and did not handle many edge cases. using stow, this
process is improved.

linking is now handled in deploy#_link_topics(). this allows us to not
have to pass the topic manually from topic/_init#link() infact, the link
method is deprecated and is replaced with variable TARGET which defines
the argument for `-t` (--target) flag in stow.

the target is checked for existance otherwise it created before stow is
invoked.