base16: simplify to just dark and light themes
I only ever use two themes, tomorrow-night for dark and solarized-light
for light. The base16-shell has a vast number of themes (which are
mostly unused by me). The loading script of the project is not that
great either, it exports a lot of variables every time a new shell is
spawned due to which fish startup time takes a hit. This commit
simplifies the base16 topic:

1) tomorrow-night and solarized-light are tracked under the topic
directly
2) fish/functions/colors simply accepts two args: `dark` or `light` and
sources the corresponding theme.
3) fish/completions/colors is removed since there are only 2 themes now