vim: flip the search highlights
It's not easy to spot the search matches which are currently not under
the cursor with a simple color highlight. This commit flips the
highlights such that the inactive matches get the colors and underline
highlight whereas the match under the cursor justs get the color. This
makes all the search matches stand out more.