tmux: fix prefix on by default issue
I ran into a weird behaviour with tmux where in, if I come back to my
tmux session after some time, pressing a key did unexpected things. I
realized that the issue was that the prefix key of tmux was already
pressed so any key pressed, if a tmux binding executed that tmux
command. I suspect the issue was because I removed the
vim-tmux-navigator commands in a917893.

After adding the keybindings back and monitoring tmux behavior for a few
days, I am certain that this infact was the issue.