[vim]: fix highlight override by colorscheme
This commit fixes the issue where my personal highlights were getting
overridden with a change in the colorscheme thus requiring a full vim
restart. The fix is based on this gist
(https://gist.github.com/romainl/379904f91fa40533175dfaec4c833f2f)
and requires the highlights to be reapplied whenever a new colorscheme
is activated with the help of autocmds. The highlights are wrapped in a
function for brevity.