vim: create custom statusline
Taking the plunge again to try and create my custom statusline. This
time, taking my time and sifting through wincent/wincent code and using
this statusline as a starting point.

With this commit, we see the basic statusline in action however the
colors and styles are missing since we require wincent/pinnacle to do
the heavy lifting for this.