[vim]: disable tex-conceal
We don't need this plugin since concealing characters in tex files makes
it super hard to edit! Also, it doesn't really do a great job of
concealing everything (at least for the documents I work on) so we end
up with this half concealed and the rest not, terribly confusing
document.