[fish]: init fish-projectionist
This feature is inspired by the wonderful plugin
[vim-projectionist](https://github.com/tpope/vim-projectionist) by Tim
Pope. The idea is to simplify visiting 'alternate' directories, here is
an example:

I keep all my documents, lectures, notes, pdfs in ~/Documents and all
software projects under ~/code. Let's say I am working on reviewing
lectures in ~/Documents/deep-learning. Once I am done, I need to start
working on the lab assignment which is under ~/code/deep-learning. Now,
my typical workflow is to open a tmux split and cd into
~/code/deep-learning. fish-projectionist makes things easier with a set
of functions namely `acd`, `scd` and `vcd` here is what they do:

1. `acd` checks if an alternate directory exists for the current
   directory you are in and takes you there.
2. `scd` does the same but opens the alternate directory in a new tmux
   split (if you are in a tmux session).
3. `vcd` same as `scd` but in a vertical tmux split.

The alternate directories are defined in a .projections.json file placed
in your fish config dir. Here is an example:

{
  "": {
    "alternate": "dotfiles"
  },
  "Documents/masters/year-2/period-1/machine-learning-1": {
    "alternate": "code/machine-learning-1"
  },
  "code/machine-learning-1": {
    "alternate": "Documents/masters/year-2/period-1/machine-learning-1"
  },
  "dotfiles": {
    "alternate": ""
  }
}

Note that the paths need to be relative to $HOME and to specify the $HOME
dir, an empty key is used.

The json file is parsed my a simple python script which by default is
assumed to be present under $XDG_DATA_BIN dir.