[git]: add git-submodule-delete
git does not have a convenient way of removing submodules and I always
need to look it up. The submodule-delete command does some preliminary
checks and warns the user of the potential side effects of the command
(which is really destructive if issued for the incorrect directory)
before executing. Ideally, we would do a check to make sure that the
provided path is in fact a submodule, potential solution would be to use
`git submodule status` and `sed` to generate a list of submodules to
check against.