[vim]: rename gf and gF to open in splits
When I use gf I almost always want it to open in a new split. Vim's built in
binding to open in split (<C-w><C-f>) is clunky and there is no option for
opening in vertical split. I have not had a usecase for gF (same as gf but
also jumps to a specific linenumber). So, we remap gf to open in split and gF
to open in vertical split. The default <C-w>gf for opening in a new tab is
fine since I rarely need it.