[vim]: give <Tab> to ultisnips when in a snippet.
The previous config (from wincent) was doing thing smartly with
autocmds where it would try to set up <Tab> and <S-Tab> to either jump
to next snippet tabstop or cycle the completion menu. However, this
stopped working with the latest deoplete and/or Ultisnips version
(perhaps an internal change?). I don't want to spend time mucking
through these package's source code so this commit simply removes this
functionality. The result is that now <Tab> and <S-Tab> is given to
ultisnips if we are in a snippet, else it is given to deoplete (ie. the
completion menu).