[fzf,vim]: fzf custom command to navigate notes
Add :WikiFiles command which wraps fzf#vim#files(), calling it directly
on the notes directory (~/vimwiki) with preview enabled. :WikiFiles! is
mapped to <Leader><C-p> for convenience (the bang version of the command
opens fzf in full screen).

Additional cosmetic changes (such as
disabling indentline, number and relativenumber in fzf buffers) were
required to make the fzf buffer look better.

Finally bat is added as a dependency for fzf since it uses it internally
in it's preview script.