[vim,zsh]: add base16 (again)
After not having the ability to switch terminal colors on the fly and
having to worry about syncing terminal and vim colors (which is
a nightmare), I realized that I need this plugin. I like the ability to
switch colors on the fly, without it I spent more time wishing I had
this functionality!

I realize that this limits my config to "smart" or not "dumb" terminal
emulators (such as emacs) but I am okay with this since I have finally
made a full time commitment to unix.