[feature] zsh: init function z
This is the first of many commit. It introduces a new zsh autoloaded
function named `z'. This function will be the main entry point to my
knowledge management system using standard unix tools. For now, it
simply enables me to quickly capture tasks and ideas. Without any
arguments, the function open my default capture file in (neo)vim, goes
to the end of the file and puts me in insert mode so I can start typing.