[feature] zsh: add agenda subcommand to z
The z command now accepts a `agenda' subcommand. The subcommand accepts
two flags namely, `--state|-s' and `--tag|-t'. With the `--state' flag,
headers matching the provided state are printed to stdin. With the
`--tag' flag, headers matching the provided tag are printed to stdin.

The command expects ripgrep (rg executable) to be available. No fault
handling is done at the moment.