[feature] zsh: init z subcommand new
The new subcommand can be used to create new note files. It takes two
flags `--tag | -t' to specify a tag for the file and `--name | -n' to
specify a string as name for the file. If neither are provided, then
a timestamp is used to name the file. The command also insert a YAML
front matter to the file.