[feature] zsh,z: capture improvements
The capture subcommand now support two new flags `--state | -s' and
`--tag | -t'. With the -s flag, we can specify a state for the header
(TODO, IDEA, etc) and with the -t a tag. Without these flags, a markdown
level 1 header is inserted at the end of the capture file and vim starts
on this line.

If the "IDEA" state is passed then a timestamp is additionally added
after the state since this is how I usually write ideas.