[feature] zsh,z: support file as argument
This is the first of many commits to support passing files as arguments
to the z command. This commit adds support to the main z command only
(not it's subcommands). If a filename is passed to z and it exists in
the NOTESDIR then the file is opened in vim, else an error is
thrown. I do not support editing any arbitrary files to promote usage of
the new command which keeps the file naming consistent.

Additionally, preliminary completion for z is also introduced in this
commit. For the time being, it presents the contents of NOTESDIR as
completion candidates. In the future I want to add support for the
subcommands and perhaps also limit the candidates to NOTESFORMAT files
only.