[feature]: zsh,z: support custom templates for new
The new subcommand now supports custom templates. `z' looks for
a .templates directory under NOTESDIR. If it contains a file with the
same name as the tag passed, then the contents of the template file is
inserted in the new notes file. Else, the default template is used.

The main motivation to do this was because paper notes are different for
me. They have a very specific set of YAML metadata and headers.