[feature] zsh,z: support line number visit
The z command now supports a special syntax for visiting the specified
note file at a specific line number. If the specified filename is in the
format `name.md:10' then z will open name.md at line 10. This is useful
for instance, for checking the open tasks with `z grep -s todo` and then
`z inbox.md:10` to view a specific task in more detail.