[feature] zsh,z: new tags format
I was previously using pandoc's format for html class and id to specify
the z tags. This in principle is wrong since they are completely
different aspects of the file. Another problem is that it prevents me
from using vim-pandoc and pandocs' intended functionality.

This commit changes the z tags format to what org uses, ie. the tags are
wrapped in colons (:).