[feature] zsh,z: use .ignore
Since inbox.md goes through a lot of change everyday, I don't like to
track it in git. This is a problem however since I want it searched with
rg. The previous solution was the add the `--no-ignore` flag however
this is not ideal.

This commit drops this flag and instead uses the .ignore file in the
NOTESDIR to override .gitignore.