[feature] zsh,z: print tags
I use tags in two ways:
+ I tag note files names
+ I tag the headers within notes files

Although I use a standard set of tags and try to keep the number of
unique tags to a minimum, I can't always remember the tags by heart.

This commit introduces the `--show-tags` flag on find and grep which
show the unique tags used along with their counts. It's important to
remember that they both work within their scopes, so grep looks for tags
within files and find looks for tags on the file names. The count is
especially useful as it gives an overview of the types of tags/topics
I write about.