[feature] zsh: init function p
This is a custom zsh function to make publishing to the internet
easier. This commit introduces the function p as a zsh autoloaded
function. It supports a single command `published` which lists the files
that are currently published (ie. a html file in the docs/ dir with the
same name exists). With the `--invert -v` flag it prints the files which
are not yet published.