[refactor] zsh,z: centralize pushd
pushd does not permit duplicates in it's stack. Thus if we invoke any
z commands from within NOTESDIR then the subsequent popd throws an
error. This commit refactors the script such that all calls to pushd are
localized to one location. The PWD is checked and only if we are not in
NOTESDIR then pushd is called. Using a flag, popd is called only if we
called pushd earlier.