[feature] zsh,z: pretty print grep results
It's cumbersome to quickly select just the file name in tmux since it
operates based on whitespace but the existing output of z grep does not
have any whitespace around the file name. This commit makes the output
of z grep a bit nicer so that the elements of the output are more
distinct from one another, and separated by whitespace.