[fix] zsh,z: general searching with grep
In 635d8ae I accidentally removed the ability to perform general
searches, this commit brings this ability back. After parsing flags, if
any arguments remain, we assume that we want to perform a directed
search. The query building phase is thus skipped entirely and GREPPRG is
invoked with the provided arguments. To pass additional arguments to
GREPPRG directly, place them after the query (else the grep command will
assume it is intended for itself and throw an error since it won't
recognise it).