[feature] zsh,z: selective agenda
The prior implementation of `agenda` searched all notes files. This
commit introduces the ability to limit the agenda to certain types and
projects. The agenda command now supports the `--type` and `--project`
flags. If any flags are passed, the output from the `find` command is
used for printing the agenda.