[feature] vim: init goyo.vim
The goyo.vim plugin provides distraction free writing in vim. However,
for the longest time I could not make it work with my custom statusline.

This commit makes the necessary changes to make goyo work nicely with my
existing statusline setup. The solution was to use the goyo_enter() and
goyo_leave() functions to turn the statusline off while entering goyo
and then turning it back on on exiting goyo.