refactor: simplify vim statusline & colors
I no longer use the base16-* projects for colorschemes, instead relying
on kitty's built-in themes and vim's default colorscheme which reflects
the terminal colors pretty well. In this commit I remove the
functionality of matching the vim colorscheme to that of my terminal.

I also simplify the statusline which no longer contains the arrow heads
on the left & right ends which indicated the buffer modification status
using color.