refactor: dont publish from docs/
This idea was short lived. I quickly realised that keeping blogposts
within the projects is a bad idea because these projects may be open
source in which case, I am giving full read+write access to my material
to others (which I think I dont want to do most of the time).