feat: init minimal dotfiles
After years of customisations and tweaking, I find that my dotfiles are
not portable at all. In this commit, I introduce the min/ dir which
contains a minimal set of dotfiles designed to be portable on all *nix
systems.

The files within min/ are organised such that I can can leverage stow's
defaults and run the following command to stow all packages:

    stow -v -t $HOME -S min