feat: init wezterm config
Trying out wezterm <https://wezfurlong.org/wezterm/>. So far the
experience has been smooth & pleasant. The terminal emulator comes
bundled with useful fonts and the text out-of-the-box is extremely
legible.

The terminal is also highly customisable and not as binding as Kitty. It
provides the perfect balance between iTerm2 & kitty. It also packs
useful features from terminal multiplexers (like tmux) such as the
ability to display the current date & time in the status bar.