refactor: drop git hooks ctags
This setup does not work with cargo new command which tries to setup
a new rust project & initialises the directory with git.

The executable is also not very universal as it does not work for:
+ directories which are not version controlled and
+ python projects with virtualenv; not indexing the virtualenv in python
  projects defeats the purpose of using ctags since the code for the
  dependencies need to be indexed.

Instead, I opt for a repetitive yet more versatile approach of using
make. This gives me the flexibility to choose my file discovery
stratergy:
+ for small projects with no external dependencies, I can simply run
  ctags -R .
+ for larger projects where I want to be selective, I can opt for find
  or fd