feat(files): init minimal system config
I have opted for bash, vim and kitty as the core pieces of software
that I want to use.

The objective moving forward is the invest time and effort only to
improve the above software.

Another important aspect, is to ensure that all config is able to
operate on any *nix platforms. Thus any new piece of configuration
must be able to gracefully fallback across different os and versions
of the software.